レイアウトのコンポーネント化
